created helper.py under LocalizationManager. worked on requested changes.
